update tests for migrations and new settings
