+++ tests/const.py
@@ -129,6 +129,73 @@ FAKE_CONFIG_DATA = {
     "verify_ssl": False,
 }
 
+# Configuration with no Amazon sensors and no custom images
+FAKE_CONFIG_DATA_NO_AMAZON = {
+    "allow_external": False,
+    "custom_img": False,
+    "ups_custom_img": False,
+    "ups_custom_img_file": "custom_components/mail_and_packages/no_deliveries_ups.jpg",
+    "folder": '"INBOX"',
+    "generate_mp4": False,
+    "gif_duration": 5,
+    "host": "imap.test.email",
+    "image_name": "mail_today.gif",
+    "image_path": "custom_components/mail_and_packages/images/",
+    "image_security": True,
+    "imap_security": "SSL",
+    "imap_timeout": 30,
+    "password": "suchfakemuchpassword",
+    "port": 993,
+    "resources": [
+        "zpackages_delivered",
+        "zpackages_transit",
+        "capost_delivered",
+        "capost_delivering",
+        "capost_packages",
+        "dhl_delivered",
+        "dhl_delivering",
+        "dhl_packages",
+        "fedex_delivered",
+        "fedex_delivering",
+        "fedex_packages",
+        "hermes_delivered",
+        "hermes_delivering",
+        "mail_updated",
+        "royal_delivered",
+        "royal_delivering",
+        "ups_delivered",
+        "ups_delivering",
+        "ups_packages",
+        "usps_delivered",
+        "usps_delivering",
+        "usps_mail",
+        "usps_packages",
+        "auspost_delivered",
+        "auspost_delivering",
+        "auspost_packages",
+        "poczta_polska_delivering",
+        "poczta_polska_packages",
+        "inpost_pl_delivered",
+        "inpost_pl_delivering",
+        "inpost_pl_packages",
+        "dpd_com_pl_delivered",
+        "dpd_com_pl_delivering",
+        "dpd_com_pl_packages",
+        "gls_delivered",
+        "gls_delivering",
+        "gls_packages",
+        "walmart_delivered",
+        "walmart_exception",
+        "post_nl_delivered",
+        "post_nl_delivering",
+        "post_nl_packages",
+    ],
+    "scan_interval": 20,
+    "storage": ".storage/mail_and_packages/images",
+    "username": "user@fake.email",
+    "verify_ssl": False,
+}
+
 FAKE_CONFIG_DATA_EXTERNAL = {
     "allow_external": True,
     "amazon_days": 3,
